Rename FilterInput to FilterButton
